All Codes/storage_to_sql/requirements.txt
google-cloud-storage
mysql-connector-python
cloudevents
